+++ domain/src/main/java/com/dida/domain/usecase/main/CommentsPostIdAPI.kt
@@ -9,7 +9,7 @@ import javax.inject.Inject
 class CommentsPostIdAPI @Inject constructor(
     private val repository: MainRepository
 ){
-    suspend operator fun invoke(postId: Int) : NetworkResult<List<Comments>> {
+    suspend operator fun invoke(postId: Long) : NetworkResult<List<Comments>> {
         return repository.getCommentsPostId(postId = postId)
     }
 }
